Decouple getMentionedUsers better
Decouple getMentionedUsers into:
getMentionedUserOneTweet, which extract usernames in a single tweet and return a seperate set, instead of appending to the existing one in the previous version;

caseInsensitiveSetUnion, which unions two string set in case-insensitive manner. Note that this function is not commutative.

Moreover, add lowerCaseSet, a util function for converting all the strings in a set into lower-case version.